Mudanças na página de contas, adicionando confirmação ao excluir e alertas quando salva/edita os dados. Além disso, também foi separado a parte de habilitar os campos para edição e a edição, conforme diagrama de classes.

+++ sigecon/src/br/sigecon/daos/ContasDAO.java
@@ -98,13 +98,15 @@ public class ContasDAO {
 		try {
 			EntityTransaction transaction = entityManager.getTransaction();
 			try {
-				transaction.begin();
-				
 				Conta conta = entityManager.find(Conta.class, id);
 				
-				entityManager.remove(conta);
-				
-				transaction.commit();
+				if (conta != null) {
+					transaction.begin();
+					
+					entityManager.remove(conta);					
+
+					transaction.commit();
+				}
 			} finally {
 				if (transaction.isActive()) {
 					transaction.rollback();
